Refactor into multiple modules
Also improves CJS support
