+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@shgysk8zer0/geoutils",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "Geo utils library for JavaScript ",
   "keywords": [
     "geo",
